PHASE ONE & TWO
working on phase one and two as one big chunk cause it just makes sense to do it that way? I'm going with two components for now, the page itself and the cards.
I've mostly set up the page, and I'm working on getting the props passed to the cards properly.